backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
google-genai>=1.0
httpx>=0.28
aiosqlite>=0.20
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0
